Live use: patch and switch effects while playing, PANIC, armed next song, setlist rack (#5)
* TODO: live use review (screen awake, playback guard, live reverb and bus 3 swap, PANIC, tune chaining, setlist rack, setlist check)


* Live use: patch and switch effects while playing, PANIC, armed next song, setlist rack

- Bus-to-bus sends go through memory (dub_portal.c): every patch works while playing, without a cut
- Plate / spring, phaser / flanger and built-in inserts switch inside the DSP (dub_switch.c): no rewiring, the old tail rings out
- PANIC: Esc, or BANK LEFT + BANK RIGHT, empties the three buses while the tune plays on
- Setlist: N / P arm the next song while playing, Space drops it, R pulls up into it
- The setlist owns the rack (effects, patches, bus plugins): nothing rewires between two songs
- Setlist check: missing stems and plugins marked before the set, one-folder relink for the whole setlist
- While playing: the screen stays on; plugin changes and stem moves wait for the stop; a plain click on the waveform no longer jumps; quit, close, New and Open ask first
- 0.6.0


---------

+++ Sources/DubStemMix/Controls.swift
@@ -126,7 +126,7 @@ struct Fader: View {
                 .contentShape(Rectangle())
                 .gesture(
                     DragGesture(minimumDistance: 0).onChanged { drag in
-                        value = min(1, max(0, 1 - (drag.location.y - capHeight / 2) / travel))
+                        value = Double(min(1, max(0, 1 - (drag.location.y - capHeight / 2) / travel)))
                     }
                 )
             }
